--- src/pages/Login/index.js
@@ -1,1 +1,0 @@
-export { default as Login } from "./login";
