--- src/pages/Signup/index.js
@@ -1,1 +1,0 @@
-export { default as Signup } from "./signup";
