+++ src/pages/index.js
@@ -1,2 +1,2 @@
-export { Login } from "./Login";
-export { Signup } from "./Signup"
+export { Login } from "./login";
+export { Signup } from "./signup"
