+++ src/pages/login/index.js
@@ -1,0 +1,1 @@
+export { default as Login } from "./Login";
