+++ src/pages/signup/index.js
@@ -1,0 +1,1 @@
+export { default as Signup } from "./Signup";
